@@ -11,6 +11,10 @@
             "email": "marcos@instasent.com"
         },
         {
+            "name": "Julián Gutiérrez",
+            "email": "juliangut@gmail.com"
+        },
+        {
             "name": "Luis Hdez",
             "email": "luis.munoz.hdez@gmail.com"
         },
@@ -21,17 +25,29 @@
         }
     ],
     "require": {
+        "php": "^7.0",
         "ext-json": "*",
         "ext-pcntl": "*",
-        "symfony/polyfill-php70": "^1.0",
-        "symfony/framework-bundle": ">=2.0",
+        "ext-posix": "*",
         "chrisboulton/php-resque": "dev-master",
-        "chrisboulton/php-resque-scheduler": "dev-master"
+        "chrisboulton/php-resque-scheduler": "dev-master",
+        "symfony/framework-bundle": ">=3.4",
+        "symfony/process": ">=3.4"
+    },
+    "require-dev": {
+        "friendsofphp/php-cs-fixer": "^2.0",
+        "squizlabs/php_codesniffer": "^3.0"
     },
     "autoload": {
-        "psr-0": {
-            "Instasent\\ResqueBundle": ""
+        "psr-4": {
+            "Instasent\\ResqueBundle\\": "src/"
         }
     },
-    "target-dir": "Instasent/ResqueBundle"
+    "target-dir": "Instasent/ResqueBundle",
+    "scripts": {
+        "lint-php": "php -l src && php -l bin",
+        "lint-phpcs": "phpcs --standard=PSR2 src bin",
+        "lint-phpcs-fixer": "php-cs-fixer fix --config=.php_cs --dry-run --verbose",
+        "fix-phpcs": "php-cs-fixer fix --config=.php_cs --verbose"
+    }
 }
