@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Instasent\ResqueBundle;
 
 abstract class Job implements JobInterface
@@ -15,39 +17,61 @@ abstract class Job implements JobInterface
     public $queue = 'default';
 
     /**
-     * @var array The job args
+     * @var mixed[] The job args
      */
-    public $args = array();
+    public $args = [];
 
+    /**
+     * {@inheritdoc}
+     */
     public function getName()
     {
         return \get_class($this);
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function getQueue()
     {
         return $this->queue;
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function getArguments()
     {
         return $this->args;
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function hasArgument($arg)
     {
         return isset($this->args[$arg]);
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function getArgument($arg)
     {
         return isset($this->args[$arg]) ? $this->args[$arg] : null;
     }
 
-    public function setArgument($arg, $value) {
+    /**
+     * {@inheritdoc}
+     */
+    public function setArgument($arg, $value)
+    {
         $this->args[$arg] = $value;
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function perform()
     {
         $this->run($this->args);
@@ -55,11 +79,17 @@ abstract class Job implements JobInterface
 
     abstract public function run($args);
 
+    /**
+     * {@inheritdoc}
+     */
     public function setUp()
     {
         // noop
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function tearDown()
     {
         // noop
