@@ -1,11 +1,13 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Instasent\ResqueBundle\DependencyInjection;
 
-use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\Config\FileLocator;
-use Symfony\Component\HttpKernel\DependencyInjection\Extension;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader;
+use Symfony\Component\HttpKernel\DependencyInjection\Extension;
 
 /**
  * This is the class that loads and manages your bundle configuration.
@@ -33,7 +35,8 @@ class InstasentResqueExtension extends Extension
 
         if (!empty($config['prefix'])) {
             $container->setParameter('instasent_resque.prefix', $config['prefix']);
-            $container->getDefinition('instasent_resque.resque')->addMethodCall('setPrefix', array($config['prefix']));
+            $container->getDefinition('instasent_resque.resque')
+                ->addMethodCall('setPrefix', [$config['prefix']]);
         }
 
         if (!empty($config['worker']['root_dir'])) {
@@ -42,13 +45,16 @@ class InstasentResqueExtension extends Extension
 
         if (!empty($config['auto_retry'])) {
             if (isset($config['auto_retry'][0])) {
-                $container->getDefinition('instasent_resque.resque')->addMethodCall('setGlobalRetryStrategy', array($config['auto_retry'][0]));
+                $container->getDefinition('instasent_resque.resque')
+                    ->addMethodCall('setGlobalRetryStrategy', [$config['auto_retry'][0]]);
             } else {
                 if (isset($config['auto_retry']['default'])) {
-                    $container->getDefinition('instasent_resque.resque')->addMethodCall('setGlobalRetryStrategy', array($config['auto_retry']['default']));
+                    $container->getDefinition('instasent_resque.resque')
+                        ->addMethodCall('setGlobalRetryStrategy', [$config['auto_retry']['default']]);
                     unset($config['auto_retry']['default']);
                 }
-                $container->getDefinition('instasent_resque.resque')->addMethodCall('setJobRetryStrategy', array($config['auto_retry']));
+                $container->getDefinition('instasent_resque.resque')
+                    ->addMethodCall('setJobRetryStrategy', [$config['auto_retry']]);
             }
         }
     }
